feat: ok, now it works (im already regretting the ugly ifdefs)
removed the ringbuffer, and solved the comm situation
with two unordered_map(s) on the receiving end.

added some ifdefs for the unique_id identification
it's not essential to the project, maybe im just going to
remove the ifdefs and increase the check of the ids to something
really large. idk if this way will scale for thousands of messages,
but in reality i don't think there is a network that is that large
in the real world.